refactor(service-page): split client and server components for better separation
Move client-side logic and effects to new ServicePageClient component to enable static generation of service pages while maintaining client-side functionality
