+++ next.config.ts
@@ -1,6 +1,7 @@
 import type { NextConfig } from "next";
 
 const nextConfig: NextConfig = {
+  //   output: "export",
 };
 
 export default nextConfig;
